feat: Pass data to article component
