+++ src/Card/Card.css
@@ -3,7 +3,8 @@
   padding: 1em;
   width: 80%;
   margin-top: 1em; 
-  color: #FFF1DA
+  color: #FFF1DA;
+  text-decoration: none;
 }
 
 .title-and-img{
